+++ packages/happn-db-provider-mongo/RELEASES.md
@@ -128,6 +128,10 @@
 -----------------
 - feat: fixed mesh_gst paths in happner-2
 
-1.1.13 2023-03-04
+1.1.12 2024-03-27
+-----------------
+- feat: loki provider has new file-size compaction threshold
+
+1.1.13 2023-05-12
 -----------------
   - Mongo-based cluster membership - dependencies update