+++ packages/happner-cluster/RELEASES.md
@@ -479,8 +479,15 @@
 -----------------
   - happner-2: happn-3 auth provider updates
 
+13.0.4 2024-03-27
+-----------------
+- feat: loki provider has new file-size compaction threshold
+
 14.0.0 2023-03-04
 -----------------
   - happn-cluster membership updated to mongo-based system
   - tests now use ports auto-assigned by OS
+
+
+